optionally display STOPPING and STOPPED hosts
Previously, the environment overview page only showed hosts in the deploy
status that were not STOPPING or STOPPED. The only way to look at them is to
look in "Total" hosts.

First of all, it doesn't make sense to me to to have an agent state that shows
up in the TOTAL but isn't available in the board.

This adds an extra configuration parameter to display STOPPING and STOPPED
hosts. It is disabled by default, in case you have some reason you do not
display those states.

Related: #354