Add ability to specify custom script directory
There are cases where we would like to replicate a service under a
different environment for management and access control. When this
happens, the current setup of teletraan forces to replicate the script
directory. Adding an option where we can over ride the assumed the
teletraan script directory. If variable TELETRAAN_SCRIPT_DIR is set, it
will supersede, the ENV_NAME / default teletraan directory.