Switch executor sleep to subprocess.wait
When using time.sleep we are forced to wait for the full sleep_time, but
if we use subprocess.Popen.wait we will wait _up_ to sleep_time.
(https://docs.python.org/3/library/subprocess.html#subprocess.Popen.wait)

With PROCESS_POLL_INTERVAL defaulting to 2 seconds, switching to this
logic can save up to 2 seconds for very fast executors, which adds up
with many stages and services on a single box.

Update test cases to use full paths for shell commands as we aren't
using shell=True