Use radio buttons to separate list filter and cluster configuration  (#950)
Previously in the UI, 'Assign Public IP' checkbox also determined what type of subnets (public or private) to show in the drop down list. In other words, if 'Assign Public IP' is checked, users are only supposed to choose from a list of public subnets. This guarantee however was broken by a bug var arr = assignPublicIp == 'true' ? this.allPublic : this.allPrivate. assignPublicIp can be a boolean and thus the expression is always evaluated to false and only private subnets are listed. As a consequence, people are allowed to choose private subnet with public IP and public subnet without public IP. Although we don't think these 2 possible combinations are practical for user scenarios, we still keep them available to minimize the impacts of this change.