Fix teletraan displays incorrect AMI
Teletraan UI shows latest AMI selected, however, new instances will launch on an older AMI. When the AMI is older, the UI shows the first (latest) AMI selected, when in reality the cluster does not have this configuration. This confused users. Fix by including the current ami in the selection dropdown.