Added a buffer window for AutoPromoter (#1002)
#392 changed the way AutoPromoter handles cron schedule. It made builds possible to be promoted every time AutoPromoter runs after their schedule time. For example, if a stage's auto promotion is scheduled at 10AM everyday, a build is possible to be promoted after the first succeeding 10AM. This doesn't meet user expectation.

To minimize the change of the whole implementation, I here propose a solution to add a small buffer window only in which promotion is possible. Currently the buffer window is set to 2 minutes. Given the fact that we run the AutoPromoter every 30 seconds, 2 minutes should be long enough to process the schedule. This is also configurable so service owners can define their own buffer window size.