Fix `stage_type` handling issue when updating `EnvStage` (#1068)
# Problem
```java
if (origBean.getStage_type() != EnvType.DEFAULT && origBean.getStage_type() != environBean.getStage_type()) {
```
The if statement results in true when the original environment stage has a non-default `stage_type` and the request doesn't intend to change the `stage_type`, which is `null`. 

# Fix
When a request has no intention to change stage type, we should set `stage_type` to the current value. Otherwise the default value will be used.