Update deploy agent logging (#1258)
* wsfix

* Add S flag to curl

From the curl documentation:

    -S, --show-error
           When used with -s, --silent, it makes curl show an error message if it fails.

Add the flag to show the error in case of failures, which an be helpful
for debugging

* Update variable names

Update the variable naming to describe the type in contains

* Update checksum verification warning

When the sha1 reuqest for a build returns non-200, the verification is
skipped.

Since this happens often and is usually expected, switch to
`log.warning`.

Also, update the warning message to clarify the sha1 verification is
skipped

* Update error logging

In case of tarball extraction failures, log the exception and stack
trace. These exceptions could cause the host to become stuck. The stack
trace can help with debugging.