Log when `delivery_type`/`stage_type` out of sync (#1432)
**Summary**

Ideally, the `delivery_type` (value passed in by deployment pipeline)
would be in sync with the `stage_type`.

However, to avoid deployment failures, simply log the request when the
values are out of sync. This will allow us find the values out of sync
and update them before enforcing the values to be in sync.

**Test Plan**

The `delivery_type` value is currently not used. Will verify existing
deployments are unaffected