Ensure `delivery_type` / `stage_type` are in sync
**Summary**

Enforce that the `delivery_type` (value passed in by deployment
pipeline) needs to be in sync with the `stage_type`.

If they are not in sync, throw an error so that users can take action
and resolve the inconsistency.

**Caveats**

In case the `delivery_type` / `stage_type` is out of sync, this could
cause deployment failures.

These failures should not happen frequently and are currently expected
to be resolved manually.

**Test Plan**

Verified there have been no inconsistencies in the last week.