Address an AuthZ flaw (#1610)
Previously Teletraan looks for the request body for AuthZ resource identifier. However it has a security flaw for some endpoints. This PR address the issue by using the identifier from request path parameter for those affected resources.

## Details of the issue
Some Teletraan endpoints follows a pattern that requires the resource identifier in the path parameter, while the request body also contains the same ID field. The ID in the path parameter is used for later processing. Therefore it's possible for a request to contain 2 resource IDs and the one used for authorization can be different from the resource being changed. 

## Changes
First of all some interface change. Previously the annotation `ResourceAuthZInfo.type` expects the type of the extracted resource. Now it's the type of the input resource. This will reduce some confusion because it matches the method it annotates. 

When appropriate, a path resource extractor replaces the previously used body resource extractor. New implementations of resource extractors are added.

The most change happen in the [EnvStageBodyExtractor.java](https://github.com/pinterest/teletraan/pull/1610/files#diff-a49e40979c111009fe3e8ecb19415406d092e63d99033e04077178dd4ec45f81). This is the result of the interface change. The implementation are split into individual extractors. 

## Tests and validation
Close to 100% test coverage for this change.