Fix long task timer implementation (#1674)
The implementation of `CumulativeHistogramLongTaskTimer` in micrometer has a known [issue](https://github.com/micrometer-metrics/micrometer/issues/2744). Besides, the implementation of `writeLongTaskTimer` is not compatible with `CumulativeHistogramLongTaskTimer` and results in non-sense histogram. The +Inf bucket should be the total counts of elements in the histogram, but it's set to the active count. 

https://github.com/pinterest/teletraan/blob/cddbb726abe2377ddc3601c8928020c6739d93f8/deploy-service/universal/src/main/java/com/pinterest/teletraan/universal/metrics/micrometer/PinStatsPublisher.java#L213-L228

Because PinStats supports GaugeHistogram, this PR adopts the second recommendation in the micrometer issue. 

> Alternatively, [GaugeHistogram](https://github.com/OpenObservability/OpenMetrics/blob/main/specification/OpenMetrics.md#gaugehistogram) is exactly what we want and doesn't require the logic of adding the previous snapshot to maintain monotonicity. Unfortunately, GaugeHistogram is not available in Prometheus' text format; it's only defined in OpenMetrics.

GaugeHistogram works for Teletraan's use case because the purpose of this metrics is to allow users to detect out of SLA host launches immediately. It's not intended to provide the histogram of actual host launch durations. 

The fix is simply to inherit from `DefaultLongTaskTimer` instead of `CumulativeHistogramLongTaskTimer`.