Map errors to exceptions (#1764)
Map WebApplicationException to more detailed subclasses, this is helpful for Rodimus adoption of this HttpClient implementation. Additionally, making `callTimeout` configurable and replacing existing `readTimeout` and `connectTimeout`. It's needed by Rodimus and it's easier to control 1 then 2. 

> The call timeout spans the entire call: resolving DNS, connecting, writing the request body, server processing, and reading the response body. If the call requires redirects or retries all must complete within one timeout period.

Lastly, there is a leak in the constructor. It's now moved to the static init block.