Handle network exception in tag worker
The tag worker is used to sync host tags in the Teletraan database.
Since this task runs in the background, eventually a network error can
be hit.

When the network error is hit, handle the exception so that the tag
worker can retry on the next run. Otherwise, the deploy constraints can
end up in an unrecoverable error state