Update deploy constraint error handling (#1840)
* Handle missing env for constraint id

While debugging issues with deploy constraints, there were many
`NullPointerException`s in the logs.

Looks like there is a cleanup issue with the deploy constraints.
Log a warning when hitting this case to help clean up the log and
distinguish this case with other deploy constraint exceptions

* Handle network exception in tag worker

The tag worker is used to sync host tags in the Teletraan database.
Since this task runs in the background, eventually a network error can
be hit.

When the network error is hit, handle the exception so that the tag
worker can retry on the next run. Otherwise, the deploy constraints can
end up in an unrecoverable error state

* Add `DeployTagWorkerTest`

Test scenarios:
- missing env for deploy constraint
- synced tags
- network exception
- unknown exception