(deploy schedule UI) Support infinite cooldown + host pct (#1847)
Support infinite cooldown. This is done in a backward-compatible way by
representing an infinite cooldown as "-1"

Also, support host percentage instead of a fixed number of the hosts for
the deploy schedule. This is done in a backward-compatible way by
specifying a percentage (e.g. "3%") instead of a fixed number (e.g. "3")