Handle duplicate hosts ids in deploy constraint
The `DeployTagWorker` syncs hosts tags that are used by the deploy
constraints.

Currently, the logic fails for host capacity. This is because a host can
be in multiple groups (e.g. application, sidecar, shards, etc.), so the
call to `getHostsByEnvId` can contain duplicate Host IDs.

These duplicates were not being handled well. Switch to a set to handle
this an properly check if the host id tags are synced in the database