Update deploy schedule last session (#1853)
The UI currently attempts to show a "final" session. This "final"
session does not actually exist in the backend. Instead, in the backend,
all hosts are allowed to deploy after the last session.

To remove this confusion, remove this "final" session. Instead, the UI
will explain the behavior that occurs after the last session.